+++ src/main/java/faang/school/postservice/service/LikeService.java
@@ -15,73 +15,14 @@ import feign.FeignException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-@Service
-@RequiredArgsConstructor
-public class LikeService {
 
-    private final PostService postService;
-    private final CommentService commentService;
-    private final UserServiceClient userServiceClient;
-    private final LikeRepository likeRepository;
-    private final UserContext userContext;
-    private final LikeMapper likeMapper;
+public interface LikeService {
+    LikeDto likePost(LikeDto likeDto);
 
-    public LikeDto likePost(LikeDto likeDto) {
-        Post post = postService.getPostIfExist(likeDto.getPostId());
-        UserDto userDto = getUserFromUserService();
-        verifyPostLikeUniqueness(post.getId(), userDto.getId());
-        Like like = Like.builder()
-                .post(post)
-                .userId(userDto.getId())
-                .build();
-        return likeMapper.toDto(likeRepository.save(like));
-    }
+    LikeDto likeComment(LikeDto likeDto);
 
+    void deleteLikePost(long postId);
 
-    public LikeDto likeComment(LikeDto likeDto) {
-        Comment comment = commentService.getCommentIfExist(likeDto.getCommentId());
-        UserDto userDto = getUserFromUserService();
-        verifyCommentLikeUniqueness(comment.getId(), userDto.getId());
-        Like like = Like.builder()
-                .comment(comment)
-                .userId(userDto.getId())
-                .build();
-        return likeMapper.toDto(likeRepository.save(like));
-
-    }
-
-
-    public void deleteLikePost(long postId) {
-        UserDto userDto = getUserFromUserService();
-        likeRepository.deleteByPostIdAndUserId(postId, userDto.getId());
-    }
-
-
-    public void deleteLikeComment(long commentId) {
-        UserDto userDto = getUserFromUserService();
-        likeRepository.deleteByCommentIdAndUserId(commentId, userDto.getId());
-    }
-
-
-    private void verifyPostLikeUniqueness(long postId, long userId) {
-        if (likeRepository.findByPostIdAndUserId(postId, userId).isPresent()) {
-            throw new DataValidationException("User by id: " + userId + " has already liked the post");
-        }
-    }
-
-    private void verifyCommentLikeUniqueness(long postId, long userId) {
-        if (likeRepository.findByCommentIdAndUserId(postId, userId).isPresent()) {
-            throw new DataValidationException("User by id: " + userId + " has already liked the comment");
-        }
-    }
-
-    private UserDto getUserFromUserService() {
-        try {
-            return userServiceClient.getUserById(userContext.getUserId());
-        } catch (FeignException e) {
-            throw new EntityNotFoundException(e.getMessage());
-        }
-    }
-
+    void deleteLikeComment(long commentId);
 
 }
